Add data access layer for questions (#5)
* fix questions directory

* add question model

* add query for answers by question id

* add insert queries for questions/answers

* add helpfulness and report queries

* fix questions query to handle missing data

Using left join instead to ensure that a question is returned even if it
has no answers and an answer is included even if it has no photos. Also,
used SELECT with alias columsn to prevent overwriting of previous values
(e.g. question.id) with NULL when there are no matching records in the
join.

* fix answers query to handle missing data

* fix insert question query

* fix insert answer query

* create unified increment helpfulness

* create unified report query