Make changes for deployment (#9)
* modify db setup to work on AWS

Trying to run the init.sql file lead to errors due to insufficient
permissions to read the .csv files for the COPY FROM. Moving those files
to /tmp/data avoids this issue since no permissions are needed to read
from this directory?

* use .sql file for database startup

The qa-server no longer attempts to setup the database if there are any
issues with connection or finding the appropriate tables. Database
management is handled separately/manually. Adds the additional .env
variables to the template.

* add production start command